refactor sdkv2 set detailed diff tests to correct files
